Modify the cmbacktrace code to remove the use of the c language standard library and adapt the printing code.

+++ program/debug/cm_tracer/cm_backtrace.h
@@ -33,8 +33,8 @@
 
 void cm_backtrace_init(const char *firmware_name, const char *hardware_ver, const char *software_ver);
 void cm_backtrace_firmware_info(void);
-size_t cm_backtrace_call_stack(uint32_t *buffer, size_t size, uint32_t sp);
-void cm_backtrace_assert(uint32_t sp);
-void cm_backtrace_fault(uint32_t fault_handler_lr, uint32_t fault_handler_sp);
+unsigned int cm_backtrace_call_stack(unsigned int *buffer, unsigned int size, unsigned int sp);
+void cm_backtrace_assert(unsigned int sp);
+void cm_backtrace_fault(unsigned int fault_handler_lr, unsigned int fault_handler_sp);
 
 #endif /* _CORTEXM_BACKTRACE_H_ */
